EE-{816/817} pulling types into separate files
