+++ execution-engine/contract-ffi-as/assembly/error.ts
