+++ execution-engine/contract-ffi-as/assembly/unit.ts
@@ -1,0 +1,9 @@
+export class Unit{
+    toBytes(): Array<u8> {
+        return  new Array<u8>(0);
+    }
+
+    static fromBytes(bytes: Uint8Array): Unit{
+        return new Unit();
+    }
+}
